Move to unity with bbatsov's upstream
In an attempt to keep the core files of the prelude as equal as possible to the
original upstream, I have moved my changes to the default values over to
personal/global-settings.el.

The changes in modules/prelude-c.el have simply been removed as they were
commented out anyway.

The auto-fill hook is now in personal/global-settings.el. I might move it over
again to core/prelude-editor.el and propose it as a pull request.